SAA - Kopie/requirements.txt
# Core Web Framework
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.11.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Claude AI Integration
anthropic==0.34.2

# Financial Data & Analysis
yfinance==0.2.33
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
statsmodels==0.14.1

# Portfolio Optimization
cvxpy==1.4.1
pypfopt==1.5.5

# Data Validation
pydantic==2.5.3
python-dotenv==1.0.0

# API & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Monitoring & Logging
python-json-logger==2.0.7